// packages/prettier-plugin-apex/vendor/apex-ast-serializer/typings/jorje.d.ts
export interface BlockComment extends HiddenToken {
    "@class": "apex.jorje.parser.impl.HiddenTokens$BlockComment";
    "@id"?: string;
    "@reference"?: string;
}
// ...
export interface HiddenToken {
    "@class": "apex.jorje.parser.impl.HiddenTokens$BlockComment" | "apex.jorje.parser.impl.HiddenTokens$InlineComment";
    location: Location;
    value: string;
}
// ...
export interface Location extends Locatable {
    "@class": "apex.jorje.data.IndexLocation" | "apex.jorje.data.PositionLocation";
    column: number;
    endIndex: number;
    line: number;
    startIndex: number;
}